Clean package init file
